Allow meteor to zip the binary instead of archiver
If you don't pass a `--directory` flag to meteor, it will zip the folder
for you.

The `archiver` package does not properly zip files with a colon in them.
For example, if you zip up a meteor build with `iron:router` and then
unzip that archived file, the output folder will be `ironrouter`.

Additionally, this removes the archiver code from this, which is nice.